feat: add permissions and delete items

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -2,13 +2,13 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
-const { transport, makeANiceEmail } = require('../mail.js');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // TODO: check if they are logged in
     if (!ctx.request.userId) {
-      throw new Error('Debes estar logado para hacer esto!')
+      throw new Error('You must be logged in to do that!');
     }
 
     const item = await ctx.db.mutation.createItem(
@@ -18,18 +18,20 @@ const Mutations = {
           user: {
             connect: {
               id: ctx.request.userId,
-            }
+            },
           },
-          ...args
-        }
+          ...args,
+        },
       },
       info
     );
 
+    console.log(item);
+
     return item;
   },
-  async updateItem(parent, args, ctx, info) {
-    // first take a copy of updates
+  updateItem(parent, args, ctx, info) {
+    // first take a copy of the updates
     const updates = { ...args };
     // remove the ID from the updates
     delete updates.id;
@@ -38,8 +40,8 @@ const Mutations = {
       {
         data: updates,
         where: {
-          id: args.id
-        }
+          id: args.id,
+        },
       },
       info
     );
@@ -47,16 +49,23 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
-    // 2.Check if they own that item or hace the permissions
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+    // 2. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
-    // toLowerCase email
+    // lowercase their email
     args.email = args.email.toLowerCase();
-    // hash the password to save. When the user want to do login, compare
-    // the hash with the password he typed.
+    // hash their password
     const password = await bcrypt.hash(args.password, 10);
     // create the user in the database
     const user = await ctx.db.mutation.createUser(
@@ -64,48 +73,45 @@ const Mutations = {
         data: {
           ...args,
           password,
-          permissions: { set: ['USER'] }
-        }
+          permissions: { set: ['USER'] },
+        },
       },
       info
     );
     // create the JWT token for them
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-    // we set the JWT as a cookie on the response
+    // We set the jwt as a cookie on the response
     ctx.response.cookie('token', token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365, // One  year
+      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
     });
-    // Finally we return the user to the browser
+    // Finalllllly we return the user to the browser
     return user;
   },
-  async signin(parent, {email, password}, ctx, info) {
-    // 1. Check if there is a user with that email
-    const user = await ctx.db.query.user({ where: { email }});
-
+  async signin(parent, { email, password }, ctx, info) {
+    // 1. check if there is a user with that email
+    const user = await ctx.db.query.user({ where: { email } });
     if (!user) {
-      throw new Error(`No se ha encontrado un usuario para ${email}`);
+      throw new Error(`No such user found for email ${email}`);
     }
     // 2. Check if their password is correct
     const valid = await bcrypt.compare(password, user.password);
-
     if (!valid) {
-      throw new Error(`La contraseña no es correcta.`);
+      throw new Error('Invalid Password!');
     }
-    // 3. Generate the jwt token
+    // 3. generate the JWT Token
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-    // 4. Set the cookie wit the token
+    // 4. Set the cookie with the token
     ctx.response.cookie('token', token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365, // One  year
+      maxAge: 1000 * 60 * 60 * 24 * 365,
     });
     // 5. Return the user
     return user;
   },
   signout(parent, args, ctx, info) {
     ctx.response.clearCookie('token');
-
-    return { message: 'Adios!' };
+    return { message: 'Goodbye!' };
   },
   async requestReset(parent, args, ctx, info) {
     // 1. Check if this is a real user
@@ -122,9 +128,8 @@ const Mutations = {
       data: { resetToken, resetTokenExpiry },
     });
     // 3. Email them that reset token
-     // 3. Email them that reset token
-     const mailRes = await transport.sendMail({
-      from: 'luismanuel.fernandez.camacho@gmail.com',
+    const mailRes = await transport.sendMail({
+      from: 'wes@wesbos.com',
       to: user.email,
       subject: 'Your Password Reset Token',
       html: makeANiceEmail(`Your Password Reset Token is here!
@@ -133,15 +138,16 @@ const Mutations = {
         .FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`),
     });
 
+    // 4. Return the message
     return { message: 'Thanks!' };
   },
   async resetPassword(parent, args, ctx, info) {
-    // 1. Check if the password match
+    // 1. check if the passwords match
     if (args.password !== args.confirmPassword) {
-      throw new Error('Tus contraseñas no coinciden');
+      throw new Error("Yo Passwords don't match!");
     }
-    // 2. Check if it's a legit reset token
-    // 3. Check if it's not expired
+    // 2. check if its a legit reset token
+    // 3. Check if its expired
     const [user] = await ctx.db.query.users({
       where: {
         resetToken: args.resetToken,
@@ -149,9 +155,9 @@ const Mutations = {
       },
     });
     if (!user) {
-      throw new Error('El token es inválido o ha expirado.');
+      throw new Error('This token is either invalid or expired!');
     }
-    // 4. Hash the new password
+    // 4. Hash their new password
     const password = await bcrypt.hash(args.password, 10);
     // 5. Save the new password to the user and remove old resetToken fields
     const updatedUser = await ctx.db.mutation.updateUser({
@@ -164,14 +170,45 @@ const Mutations = {
     });
     // 6. Generate JWT
     const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
-    // 7. Set JWT cookie
+    // 7. Set the JWT cookie
     ctx.response.cookie('token', token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365, // One  year
+      maxAge: 1000 * 60 * 60 * 24 * 365,
     });
-    // 8. Return the new user
+    // 8. return the new user
     return updatedUser;
-  }
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
+      },
+      info
+    );
+    // 3. Check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info
+    );
+  },
 };
 
 module.exports = Mutations;
